Move extractCookies into separate file
